Some code moved out of cycles
There was a problem that 12x column compiled into 118k lines of code. That means that it take upt to 1 mb of data.
Now we lost feature that not able user to use more cell-Nx where N > M for grid-Mx. Now it is possible but makes a small bug that will be fixed soon. Now it ony takes maximum place. It also able you to use multiple fixeds in different lines.